refactor(tools): break the tools <-> plan-executor require cycle
The orchestrator's plan-executor imported ToolCall/UnifiedSource from the
big tools.ts, which imports the executor back - a no-circular violation the
dependency-boundaries gate rejects. Move the two shared leaf types to
tools/tool-types.ts (imported by both, importing nothing back); tools.ts
re-exports them so external importers are unchanged.
